pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "datasynth"
version = "0.1.0"
description = "Streamlit front-end framework for synthetic multi-table data generation"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "streamlit>=1.31",
  "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0",
]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["pytest"]
